pyproject.toml
[tool.poetry]
name = "weiwo"
version = "0.1.0"
description = ""
authors = ["ibroxim <ibrohimabdusalomov395@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.25.1"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.4"
asyncpg = "^0.27.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
